backend/pyproject.toml
[project]
name = "smartshelf-api"
version = "0.1.0"
description = "Smart Shelf V2 — FastAPI backend with VLM-powered product scanning"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "motor>=3.7.0",
    "pymongo>=4.10.0",
    "pyzbar>=0.1.9",
    "Pillow>=11.0.0",
    "transformers>=4.48.0",
    "torch>=2.5.0",
    "qwen-vl-utils>=0.0.8",
    "accelerate>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
